requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-rest==0.8.7
djangorestframework==3.13.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
psycopg2-binary==2.9.1
